refactor: extract sandbox shell syntax
